# specs/001-fullstack-web-app/contracts/api.yaml
openapi: 3.0.3
info:
  title: Todo Web Application API
  description: RESTful API for the Phase II Full-Stack Todo Application
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.yourdomain.com
    description: Production

tags:
  - name: Authentication
    description: User signup, signin, and signout
  - name: Tasks
    description: Task CRUD operations
  - name: Health
    description: Health check endpoint

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Returns server health status
      operationId: healthCheck
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /api/auth/signup:
    post:
      tags:
        - Authentication
      summary: Create new account
      description: Register a new user with email and password
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: JWT token in httpOnly cookie
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/signin:
    post:
      tags:
        - Authentication
      summary: Sign in
      description: Authenticate with email and password
      operationId: signin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Signed in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: JWT token in httpOnly cookie
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/auth/signout:
    post:
      tags:
        - Authentication
      summary: Sign out
      description: Clear session and invalidate JWT
      operationId: signout
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Signed out successfully
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Cleared cookie
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks:
    get:
      tags:
        - Tasks
      summary: List all tasks
      description: Get all tasks for the authenticated user
      operationId: listTasks
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User ID mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Tasks
      summary: Create task
      description: Create a new task for the authenticated user
      operationId: createTask
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User ID mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks/{id}:
    get:
      tags:
        - Tasks
      summary: Get task
      description: Get a single task by ID
      operationId: getTask
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User ID mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Tasks
      summary: Update task
      description: Update task title and/or description
      operationId: updateTask
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User ID mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Tasks
      summary: Delete task
      description: Permanently delete a task
      operationId: deleteTask
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      responses:
        '204':
          description: Task deleted
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User ID mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/{user_id}/tasks/{id}/complete:
    patch:
      tags:
        - Tasks
      summary: Toggle completion
      description: Mark task as complete or incomplete
      operationId: toggleComplete
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/TaskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskComplete'
      responses:
        '200':
          description: Completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User ID mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT token stored in httpOnly cookie

  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      description: User ID (must match authenticated user)
      schema:
        type: string

    TaskId:
      name: id
      in: path
      required: true
      description: Task ID
      schema:
        type: integer

  schemas:
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: sarah@example.com
        password:
          type: string
          minLength: 8
          example: password123

    AuthResponse:
      type: object
      properties:
        user:
          type: object
          properties:
            id:
              type: string
              example: user_123abc
            email:
              type: string
              example: sarah@example.com
            name:
              type: string
              nullable: true
              example: Sarah

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: Milk, eggs, bread

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: Buy groceries and fruits
        description:
          type: string
          maxLength: 1000
          nullable: true
          example: Milk, eggs, bread, apples, bananas

    TaskComplete:
      type: object
      required:
        - completed
      properties:
        completed:
          type: boolean
          example: true

    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: string
          example: user_123abc
        title:
          type: string
          example: Buy groceries
        description:
          type: string
          nullable: true
          example: Milk, eggs, bread
        completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: '2025-12-29T10:30:00Z'
        updated_at:
          type: string
          format: date-time
          example: '2025-12-29T10:30:00Z'

    TaskListResponse:
      type: object
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'

    ErrorResponse:
      type: object
      required:
        - error
        - detail
        - code
      properties:
        error:
          type: string
          example: ValidationError
        detail:
          type: string
          example: Title must be between 1 and 200 characters
        code:
          type: string
          example: INVALID_TITLE_LENGTH